Extract item pipeline from scrapy_wrapper
